Поменял шаг настройки времени с 0.5 сек до 1 сек

+++ app/src/main/java/ru/s4nchez/chords/presentation/view/settings/SettingsView.kt
@@ -1,6 +1,6 @@
 package ru.s4nchez.chords.presentation.view.settings
 
 interface SettingsView {
-    fun showChordTimeInSeconds(seconds: Double)
+    fun showChordTimeInSeconds(seconds: Int)
     fun setSeekBarProgress(progress: Int)
 }
